requirements.txt
Django
psycopg2>=2.8
django-phonenumber-field[phonenumbers]